LayoutTests/fast/css/parsing-css-matches-5-expected.txt

Test the parsing of :matches(selectorList) for querySelector and style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


1 invalid selectors
PASS document.querySelector(":matches()") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(    )") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(,)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(.123class)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(#123id)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches([])") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches())") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches({})") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(})") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(:not()") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(:matches()") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS document.querySelector(":matches(:nth-child(2n+1 of)") threw exception Error: SyntaxError: DOM Exception 12.
PASS document.getElementById('style-container').sheet.cssRules.length is 0
PASS successfullyParsed is true

TEST COMPLETE
